Leah - Updated the main with random number functions and difficulty spacing.
